Distances

GeneticVariation.jl allows one to compute a range of distance measures that are relevant to the analysis of genetic variation.

MASH distances

GeneticVariation.MASH — Type

MASH distances, based on MinHash sketches of genome sequences provide rapid genome-scale sequence comparisons when sequence distance (not specific mutations) are all that's required.

A MinHash sketch is made by taking the s smallest hash values for kmers of length k for a given sequence. The genome distance for two genomes is then essentially the Jaccard index of the minhashes, with some additional modification to account for the size of the kmers used.

You can generate a MinHash sketch using the minhash() function in our BiologicalSequences.jl package.

using BiologicalSequences

seq1 = dna"AAATAAGGCACAACTATGCAT"
sketch1 = minhash(seq, 5, 10)

Then, if you have MinHash sketches with the same parameters for two sequences, you can determine the MASH distance between them.

seq2 = dna"AATTAACGCACGGACTGCGGTAAT"
sketch2 = minhash(seq, 5, 10)

using GeneticVariation

distance(MASH, sketch1, sketch2)
# or use the convenience function
mash(sketch1, sketch2)

For more information on what size kmers and what size sketches are appropriate for your use-case, see Odnov et. al. in Genome Biology.